\name{expit}
\alias{expit}
\title{Perform expit transformation.}
\description{Applies expit transformation to one or more real numbers, which is the inverse of the logit transformation applied to one or more probabilities.}
\usage{
expit(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Any numeric value, vector, or array.}
}
\value{A value, vector, or array matching the dimensions of x and with elements, y, with range 0 < y < 1.}
\author{
Quresh S. Latif
}

\seealso{logit}
