\name{plotNestDens}
\alias{plotNestDens}
\title{
  Plot sample density along HSI gradient
}
\description{
  Generates plot of observed (nest) density in moving window bins and suitability classes along habitat suitability index (HSI) gradient.
}
\usage{
  plotNestDens(dat.plot, dat.class, nests, thresholds, binPntSize = 2, classPntSize = 5,
                         axisLabSize = 20, tickLabSize = 15, classLabSize = 5, lab.params, Dens.adj = T,
                         BS = F, ylabel = "Observed density (nests per 100 ha)")
}
\arguments{
  \item{dat.plot}{Data frame with one row for each moving window bin. Ideally, use output from 'calcBinDensities' function.}
  \item{dat.class}{Data frame with one row for each suitability class. Ideally, use output from 'calcClassDensities' function.}
  \item{sampleHSIs}{Numeric vector with HSI values for sample observations (e.g., nest locations; used for rug plot)}
  \item{thresholds}{Numeric vector of HSI breakpoints defining suitability classes}
  \item{binPntSize}{Size of points representing bin densities in plot}
  \item{classPntSize}{Size of points representing class densities in plot}
  \item{axisLabSize}{Size of axis titles}
  \item{tickLabSize}{Size of tick labels}
  \item{classLabSize}{Size of suitability class labels (low, medium, and high)}
  \item{lab.params}{List of positions and sizes of additional plot labels - "maxSSS" labeling the maxSSS threshold, and suitability class labels "Low", "Moderate", and "High". Names of list elements should be the labels "maxSSS", "Low", "Moderate", and "High". Each element is a numeric vector containing x-position, y-position, and angle of label in that order.}
  \item{BS}{If true, dat.class must include confidence limits (e.g., those generated by the 'HSIClassDensityBS' function), add error bars will be plotted depicting these.}
  \item{ylabel}{Character string - label for y-axis}
}
\details{Generates plot (using ggplot2 package) showing sample (nest) densities along HSI gradient for publication and assessing goodness-of-fit.}
\author{
  Quresh S. Latif, Rocky Mountain Research Station, U.S. Forest Service
}

\examples{
  ## The function is currently defined as
  function(dat.plot, dat.class, sampleHSIs, thresholds, binPntSize = 2, classPntSize = 5,
                         axisLabSize = 20, tickLabSize = 15, classLabSize = 5, lab.params,
                         BS = F, ylabel = "Observed density (nests per 100 ha)") {
  theme_set(theme_bw())
  plt <- ggplot(dat.plot, aes(HSI.md, Density_adj)) +
    geom_point(size = binPntSize, shape = 1) +
    geom_point(data = dat.class, aes(x = HSI.md, y = Density_adj), size = classPntSize, shape = 16) +
    geom_rug(data = sampleHSIs, aes(x = HSI_WHWO, y = NULL), colour = "black", alpha = 0.3, size = 1) +
    geom_vline(xintercept = thresholds[1], linetype = "dashed") +
    geom_vline(xintercept = thresholds[2], linetype = "dashed") +
    ylab(ylabel) + xlab("Habitat Suitability Index (HSI)") +
    theme(axis.text.x=element_text(size=tickLabSize)) +
    theme(axis.text.y=element_text(size=tickLabSize)) +
    theme(axis.title.x=element_text(size=axisLabSize)) +
    theme(axis.title.y=element_text(size=axisLabSize)) +
    annotate("text", x = lab.params[["maxSSS"]][1], y = lab.params[["maxSSS"]][2], label = "maxSSS", angle = 90) +
    annotate("text", x = lab.params[["Low"]][1], y = lab.params[["Low"]][2], label = "Low", size = classLabSize) +
    annotate("text", x = lab.params[["Moderate"]][1], y = lab.params[["Moderate"]][2], label = "Moderate",
             size = classLabSize, angle = lab.params[["Moderate"]][3]) +
    annotate("text", x = lab.params[["High"]][1], y = lab.params[["High"]][2], label = "High", size = classLabSize)
  
  if(BS == T) plt <- plt +
    geom_errorbar(data = dat.class, aes(x = HSI.md, ymin = Dens95lo, ymax = Dens95hi), size = 1, width = 0.05)
  
  return(plt)
}
}
\keyword{models}% use one of  RShowDoc("KEYWORDS")
\keyword{HSI}
\keyword{Resource Selection Function}
\keyword{habitat}
\keyword{model evaluation}
\keyword{density}
