\name{WLR_fit}
\alias{WLR_fit}
\title{
Fit weighted logistic regression model to use-nonuse and habitat data.
}
\description{
Fits weighted logistic regression to binary data where 1 = habitat use observation and 0 = non-use observation. Weights are set so that the sum of weights for 0s will equal the number of 1s, and sample size will be twice the number of 1s. This type of model was used to develop habitat suitability models using nest location data for disturbance-associated woodpeckers (Russell et al. 2007, Latif et al. 2016).
}
\usage{
WLR_fit(dat, formula, Obs = "Nest")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
Data frame with one column indicating which observations are use (1) versus non-use (0), and additional columns containing habitat covariates to be used as predictors in the model.
}
  \item{mod}{
A character string containing the right hand side of the desired model, i.e., of the form "x1+x2+...+xK", where 'xK' = the last covariate in the model.
}
  \item{Obs}{
The name of the column in 'dat' indicating use-nonuse, i.e., y in the formula y~x.
}
}
\details{
The model is fit using glm() with family = "binomial" and weights specified as described above.
}
\value{
Model object generated by glm()
}
\references{
Latif, Q. S., V. A. Saab, J. P. Hollenbeck, and J. G. Dudley. 2016. Transferability of habitat suitability models for nesting woodpeckers associated with wildfire. The Condor 118:766-790.

Russell, R. E., V. A. Saab, and J. G. Dudley. 2007. Habitat-suitability models for cavity-nesting birds in a postfire landscape. Journal of Wildlife Management 71:2600-2611.
}
\author{
Quresh S. Latif, Rocky Mountain Research Station, U.S. Forest Service
}

\seealso{
glm, predict.glm
}
\examples{
## The function is currently defined as
function(dat, mod, Obs = "Nest") {
  form <- paste0(Obs, "~", mod)
  w <- rep(1, nrow(dat))
  w[which(dat[, Obs] == 0)] <- sum(dat[, Obs] == 1) / sum(dat[, Obs] == 0)
  dat$w <- w
  mod <- glm(form, family = binomial, weights = w, data = dat)
  return(mod)
}
}
\keyword{models}% use one of  RShowDoc("KEYWORDS")
\keyword{glm}
\keyword{woodpecker}
\keyword{forest disturbance}
\keyword{nesting}
\keyword{dry forest management}
