\name{WoodpeckerHSI-package}
\alias{WoodpeckerHSI-package}
\alias{WoodpeckerHSI}
\docType{package}
\title{Woodpecker HSI functions}
\version{1.01}
\description{Functions and data for applying habitat suitability models for disturbance associated woodpeckers as described in various publications by Rocky Mountain Research Station scientists.}
\author{Quresh Latif}
\references{
Hollenbeck, J. P., V. A. Saab, and R. W. Frenzel. 2011. Habitat suitability and nest survival of White-headed Woodpeckers in unburned forests of Oregon. Journal of Wildlife Management 75:1061-1071.

Latif, Q. S., V. A. Saab, J. G. Dudley, and J. P. Hollenbeck. 2013. Ensemble modeling to predict habitat suitability for a large-scale disturbance specialist. Ecology and Evolution 3:4348-4364.

Latif, Q. S., V. A. Saab, K. Mellen-Mclean, and J. G. Dudley. 2015. Evaluating habitat suitability models for nesting white-headed woodpeckers in unburned forest. The Journal of Wildlife Management 79:263-273.

Latif, Q. S., V. A. Saab, J. P. Hollenbeck, and J. G. Dudley. 2016. Transferability of habitat suitability models for nesting woodpeckers associated with wildfire. The Condor 118:766-790.

Russell, R. E., V. A. Saab, and J. G. Dudley. 2007. Habitat-suitability models for cavity-nesting birds in a postfire landscape. Journal of Wildlife Management 71:2600-2611.

Wightman, C. S., V. A. Saab, C. Forristal, K. Mellen-McLean, and A. Markus. 2010. White-headed Woodpecker nesting ecology after wildfire. Journal of Wildlife Management 74:1098-1106.
}
\keyword{package}
\keyword{disturbance-associated woodpeckers}
\keyword{dry conifer forest management}
