\name{plotDens}
\alias{plotDens}
\title{
  Plot sample density along HSI gradient
}
\description{
  Generates plot of observed (nest) density in moving window bins and suitability classes along habitat suitability index (HSI) gradient.
}
\usage{
  plotDens(dat.plot, dat.class, sampleHSIs, thresholds, binPntSize = 2, classPntSize = 5,
                         axisLabSize = 20, tickLabSize = 15,
                         BS = F, ylabel = "Observed density (nests per 100 ha)",
                         xlabel = "Habitat Suitability Index (HSI)")
}
\arguments{
  \item{dat.plot}{Data frame with one row for each moving window bin. Ideally, use output from 'calcBinDensities' function.}
  \item{dat.class}{Data frame with one row for each suitability class. Ideally, use output from 'calcClassDensities' function.}
  \item{sampleHSIs}{Numeric vector with HSI values for sample observations (e.g., nest locations; used for rug plot)}
  \item{thresholds}{Numeric vector of HSI breakpoints defining suitability classes}
  \item{binPntSize}{Size of points representing bin densities in plot}
  \item{classPntSize}{Size of points representing class densities in plot}
  \item{axisLabSize}{Size of axis titles}
  \item{tickLabSize}{Size of tick labels}
  \item{BS}{If true, dat.class must include confidence limits (e.g., those generated by the 'HSIClassDensityBS' function), add error bars will be plotted depicting these.}
  \item{ylabel}{Character string - label for y-axis}
  \item{xlabel}{Character string - label for x-axis}
}
\details{Generates plot (using ggplot2 package) showing sample (nest) densities along HSI gradient for publication and assessing goodness-of-fit.}
\author{
  Quresh S. Latif, Rocky Mountain Research Station, U.S. Forest Service
}

\keyword{models}% use one of  RShowDoc("KEYWORDS")
\keyword{HSI}
\keyword{Resource Selection Function}
\keyword{habitat}
\keyword{model evaluation}
\keyword{density}
